innoveraibot-vision-based-summary-service/requirements.txt
azure-functions==1.18.0
azure-storage-blob==12.19.1
azure-storage-queue==12.9.0
PyJWT==2.8.0
setuptools
wheel

# Langchain
#langchain==0.1.11
langchain==0.1.16
langchain-openai==0.0.8
#openai==1.13.3
openai==1.23.6
pymongo==4.6.2
tiktoken==0.6.0
pypdfium2==4.28.0
flashrank==0.2.0
simsimd==3.7.7
pillow
# Langchain - Summarization
langchain-core
tenacity==8.3.0
mongoengine==0.27.0
pymupdf